// src/app/modules/core/direct-business/models/doiModel.ts
export class PartyMasterOfficeDetails {
    officeType: string;
    officeName: string;
    officeAddress: string;
    district: string;
    taluka: string;
    city: string;
    pinCode: string;
    talukaList: []
}

export class PartyMasterLocationDetails {
    locationType: string;
    locationName: string;
    locationAddress: string;
    district: string;
    taluka: string;
    city: string;
    pinCode: string;
    talukaList: []
}

export class PartyMasterBankDetails {
    bankId: string;
    bankName: string;
    branchName: string;
    branchId: string;
    ifscCode: string;
    accountNo: string;
    paymentPrefered: string;
    BranchCode: string;
    branchType: string;
    branchList: any[];
}